reorganise md converter
