+++ pyproject.toml
@@ -146,12 +146,11 @@ interop-crewai = [
 interop-langchain = ["langchain-community>=0.3.12,<1"]
 interop-pydantic-ai = ["pydantic-ai==0.0.13"]
 interop = [
-    "pyautogen[interop-crewai, interop-langchain, interop-pydantic-ai,tool_convert_markitdown,tool_convert_docling]",
+    "pyautogen[interop-crewai, interop-langchain, interop-pydantic-ai,tool_convert_markitdown]", #tool_convert_docling
 ]
 
 test-ci = [
-    "pyautogen[test,cosmosdb,interop,tool_convert_markitdown,tool_convert_docling]",
-    "pydantic>=2.0.0"
+    "pyautogen[test,cosmosdb,interop,tool_convert_markitdown]", #tool_convert_docling
 ]
